+++ ptai-ee-tools-java/ptai-cli-plugin/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/cli/Plugin.java
@@ -6,10 +6,7 @@ import org.fusesource.jansi.AnsiConsole;
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
 
-import java.io.IOException;
-import java.io.InputStream;
 import java.util.Comparator;
-import java.util.logging.LogManager;
 
 @Command (name = "java -jar ptai-cli-plugin.jar", synopsisSubcommandLabel = "COMMAND",
         mixinStandardHelpOptions = true, versionProvider = ManifestVersionProvider.class,
@@ -65,15 +62,6 @@ public class Plugin implements Runnable {
     static class SortByOrder<T extends CommandLine.Model.IOrdered> implements Comparator<T> {
         public int compare(T o1, T o2) {
             return Integer.signum(o1.order() - o2.order());
-        }
-    }
-
-    static {
-        try {
-            InputStream stream = Plugin.class.getResourceAsStream("/logging.properties");
-            LogManager.getLogManager().readConfiguration(stream);
-        } catch (IOException e) {
-            e.printStackTrace();
         }
     }
 }